+++ bin/.gitignore
@@ -1,0 +1,2 @@
+/Application.class
+/Application$1.class
