+++ src/ScoreTimer.java
@@ -16,22 +16,6 @@ public class ScoreTimer extends JFrame {
 	private JPanel contentPane;
 	
 	/**
-	 * Launch the application.
-	 */
-	public static void main(String[] args) {
-		EventQueue.invokeLater(new Runnable() {
-			public void run() {
-				try {
-					ScoreTimer frame = new ScoreTimer();
-					frame.setVisible(true);
-				} catch (Exception e) {
-					e.printStackTrace();
-				}
-			}
-		});
-	}
-	
-	/**
 	 * Create the frame.
 	 */
 	public ScoreTimer() {
